Give up on inpermanence configuration
After about a weeks worth of effort attempting to modernize this AWS
based configuration into a my flake based project.
https://christine.website/blog/paranoid-nixos-aws-2021-08-11

I have now accepted that I am currently unable to do so, primarily due
to my lack of experience around the inner workings of the
make-disk-image.nix file.

If I return to this in the future, the main challenges that i got stuck
on were to build make-disk-image into a tool that could support a non
root mountpoint. And that the inpermanence bind mounts would not be
built into the image pointing to `/mnt/root/...`.