Failed Authentik configuration
This is a mostly a tale of attempting to use badly packaged software.
Authentik seemed like a good solution at the time to attempt to build my
sso infrastructure on. But the longer i have spendt attempting to make
it work nicely as a rootless container, and finding particularly nasty
oversights by its developer (SIGTERMs did not have any handler, every
time you shut the software down, you had to SIGKILL it...), i have
decided to scrap my efforts.

This commit is here mostly for a caustionary tale, and to commit my
attempted solution to git for others to see.

If you still want to attempt running it, the primary issue were
permission related, where software wanted to edit files its not allowed
to. But also not being clear which files it would like to change. (You
will only get a vauge PermissionError).