flake.nix: Revert nixpkgs-patched due to IFD
IFD, or in its long form, Input from Derivation, is a nasty gotcha that
is easy to forget about inside of Nix.

In very short, `import nixpkgs-patched { ... }` as my package set
requires the `/nix/store/hash-nixpkgs-patched.drv` to be available.

This is not available inside of a `flake check --no-build` enviroment,
and would require the entirety of nixpkgs-patched to be realised in each
CI build. Which is quite a task.

For now, building overlays manually for each patch seems to be the best
method forwards. I may write a simple script to create these if PR
overlays become more common for me to implement.

I have attached a resource[1] about IFD for future reference on this
limitation.

1: https://fzakaria.com/2020/10/21/nix-parallelism-import-from-derivation.html